Return EXIT_SUCCESS if user.pax.flags is done after paxctl-ng -d
If the user.pax.flags field exists and we have permissions to remove
it, the first invocation of paxctl-ng -d returns 0.  But subsequently
it returns 1 because it fails to remove an xattr field that is not
there.  We make sure we return 0 if the field is gone for whatever
reason.  We only fail upon not having permissions to change the xattr
filed, or the filesystem not supporting xattrs (ENOTSUP).

Reported-by: Maxim Kammerer <mk@dee.su>

X-Gentoo-Bug: 485908
X-Gentoo-Bug-URL: https://bugs.gentoo.org/485908