misc/elf-abi: manually read little/big endian half or full words
When reading on a native system, read() transparently reads two
byte or four bytes in the correct byte order and returns the correct
value.  But if you read() a word in a big endian file from a little
endian system or vice versa, you get the wrong value.  We wrap
read() in read_endian() to compensate for this.